+++ source/features/toggle-all-things-with-alt.js
